dtt::datetime
Function days_in_month

pub const fn days_in_month(year: i32, month: u8) -> Result<u8, DateTimeError>

Helper function to determine the number of days in a given month and year.

Arguments

year - Calendar year

month - Month number (1-12)

Returns

Returns a Result containing either the number of days or a DateTimeError.

Errors

Returns a DateTimeError if the day in the month is invalid.