Struct dtt::datetime::DateTime

pub struct DateTime {
    pub datetime: PrimitiveDateTime,
    pub offset: UtcOffset,
}

A structure representing a date and time with timezone information.

Fields
datetime: PrimitiveDateTime

The date and time in UTC.

offset: UtcOffset

The timezone offset in UTC.

Implementations

impl DateTime

pub fn new() -> Self

Creates a new DateTime instance with the current date and time in UTC.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
println!("Current UTC time: {}", dt);

pub fn new_with_tz(tz: &str) -> Result<Self, DateTimeError>

Creates a new DateTime instance with the current date and time in the specified timezone.

Arguments

tz - A string slice that holds the timezone abbreviation (e.g., “UTC”, “EST”, “PST”).

Returns

Result<Self, DateTimeError> - The new DateTime instance or an error if the timezone is invalid.

Note

This function supports only fixed UTC offsets. Daylight Saving Time (DST) is not automatically adjusted. If DST adjustments are needed, they must be handled manually by choosing the correct timezone offset.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new_with_tz("EST").expect("EST is a valid timezone");
println!("Current time in EST: {}", dt);

pub fn new_with_custom_offset( hours: i8, minutes: i8, ) -> Result<Self, DateTimeError>

Creates a new DateTime instance with the current date and time in a custom timezone offset.

Arguments

hours - The hour component of the offset.

minutes - The minute component of the offset.

Returns

Result<Self, DateTimeError> - The new DateTime instance or an error if the offset is invalid.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new_with_custom_offset(5, 30).expect("Valid custom offset");
println!("Current time with custom offset: {}", dt);

pub fn from_components( year: i32, month: u8, day: u8, hour: u8, minute: u8, second: u8, offset: UtcOffset, ) -> Result<Self, DateTimeError>

Creates a DateTime instance from individual components.

Arguments

year - The year to set.

month - The month to set (1-12).

day - The day of the month to set.

hour - The hour to set (0-23).

minute - The minute to set (0-59).

second - The second to set (0-59).

offset - The UtcOffset to set.

Returns

Result<Self, DateTimeError> - A new DateTime instance created from the specified components, or an error if any component is invalid.

Examples

use dtt::datetime::DateTime;
use time::UtcOffset;

let dt = DateTime::from_components(2023, 5, 20, 15, 30, 0, UtcOffset::UTC).expect("Valid date components");
println!("Created date: {}", dt);

pub fn parse(input: &str) -> Result<Self, DateTimeError>

Parses a string into a DateTime instance.

This method supports RFC 3339 and ISO 8601 date formats.

Arguments

input - A string slice that holds the date/time to parse.

Returns

Result<Self, DateTimeError> - The parsed DateTime instance or an error if parsing fails.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::parse("2023-05-20T15:30:00Z").expect("Valid RFC 3339 date");
println!("Parsed date: {}", dt);

pub fn parse_custom_format( input: &str, format: &str, ) -> Result<Self, DateTimeError>

Parses a string into a DateTime instance using a custom format.

Arguments

input - A string slice that holds the date/time to parse.

format - The format string to use for parsing.

Returns

Result<Self, DateTimeError> - The parsed DateTime instance or an error if parsing fails.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::parse_custom_format("2023-05-20 15:30:00", "[year]-[month]-[day] [hour]:[minute]:[second]").expect("Valid custom format");
println!("Parsed date: {}", dt);

pub fn update(&self) -> Result<Self, DateTimeError>

Updates the DateTime instance to the current date and time.

Returns

Result<Self, DateTimeError> - The updated DateTime instance or an error if the update fails.

Examples

use dtt::datetime::DateTime;
use std::thread::sleep;
use std::time::Duration;

let mut dt = DateTime::new();
sleep(Duration::from_secs(1));
dt = dt.update().expect("Update should succeed");
println!("Updated time: {}", dt);

pub fn convert_to_tz(&self, new_tz: &str) -> Result<Self, DateTimeError>

Converts the DateTime instance to another timezone.

Arguments

new_tz - The timezone abbreviation to convert to.

Returns

Result<Self, DateTimeError> - A new DateTime instance in the target timezone or an error if the conversion fails.

Note

This function only supports fixed UTC offsets. If the target timezone observes DST, you will need to handle the offset change manually.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new_with_tz("EST").expect("Valid timezone");
let converted_dt = dt.convert_to_tz("PST").expect("Conversion to PST should succeed");
println!("Converted time: {}", converted_dt);

pub fn unix_timestamp(&self) -> i64

Returns the Unix timestamp of the DateTime instance.

Returns

i64 - The Unix timestamp in seconds.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
let timestamp = dt.unix_timestamp();
println!("Unix timestamp: {}", timestamp);

pub fn add_days(&self, days: i64) -> Result<Self, DateTimeError>

Adds a specified number of days to the DateTime instance.

Arguments

days - The number of days to add (can be negative for subtraction).

Returns

Result<Self, DateTimeError> - A new DateTime instance with the days added, or an error if the operation fails.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
let future_dt = dt.add_days(7).expect("Adding 7 days should succeed");
println!("Date after 7 days: {}", future_dt);

pub fn next_day(&self) -> Result<Self, DateTimeError>

Returns a new DateTime instance representing the next day.

Returns

Result<Self, DateTimeError> - A new DateTime instance for the next day, or an error if the operation fails.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
let tomorrow = dt.next_day().expect("Getting next day should succeed");
println!("Tomorrow: {}", tomorrow);

pub fn previous_day(&self) -> Result<Self, DateTimeError>

Returns a new DateTime instance representing the previous day.

Returns

Result<Self, DateTimeError> - A new DateTime instance for the previous day, or an error if the operation fails.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
let yesterday = dt.previous_day().expect("Getting previous day should succeed");
println!("Yesterday: {}", yesterday);

pub fn format(&self, format_str: &str) -> Result<String, DateTimeError>

Formats the DateTime instance according to the specified format string.

Arguments

format_str - A string slice that holds the desired format.

Returns

Result<String, DateTimeError> - The formatted date/time string or an error if formatting fails.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
let formatted = dt.format("[year]-[month]-[day]").expect("Format should be valid");
println!("Formatted date: {}", formatted);

pub fn format_rfc3339(&self) -> Result<String, DateTimeError>

Formats the DateTime instance as an RFC 3339 string.

Returns

Result<String, DateTimeError> - The formatted RFC 3339 string or an error if formatting fails.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
let formatted = dt.format_rfc3339().expect("RFC 3339 format should succeed");
println!("RFC 3339 date: {}", formatted);

pub fn format_iso8601(&self) -> Result<String, DateTimeError>

Formats the DateTime instance as an ISO 8601 string.

Returns

Result<String, DateTimeError> - The formatted ISO 8601 string or an error if formatting fails.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
let formatted = dt.format_iso8601().expect("ISO 8601 format should succeed");
println!("ISO 8601 date: {}", formatted);

pub fn duration_since(&self, other: &Self) -> Duration

Calculates the duration between this DateTime and another DateTime.

Arguments

other - The DateTime to compare with.

Returns

Duration - The duration between the two DateTime instances.

Examples

use dtt::datetime::DateTime;

let dt1 = DateTime::new();
let dt2 = dt1.add_days(1).expect("Adding 1 day should succeed");
let duration = dt1.duration_since(&dt2);
println!("Duration: {:?}", duration);

pub fn start_of_week(&self) -> Result<Self, DateTimeError>

Returns the start of the week for the DateTime instance.

Returns

Result<Self, DateTimeError> - A new DateTime instance at the start of the week.

pub fn end_of_week(&self) -> Result<Self, DateTimeError>

Returns the end of the week for the DateTime instance.

Returns

Result<Self, DateTimeError> - A new DateTime instance at the end of the week.

pub fn start_of_month(&self) -> Result<Self, DateTimeError>

Returns the start of the month for the DateTime instance.

Returns

Result<Self, DateTimeError> - A new DateTime instance at the start of the month.

pub fn end_of_month(&self) -> Result<Self, DateTimeError>

Returns the end of the month for the DateTime instance.

Returns

Result<Self, DateTimeError> - A new DateTime instance at the end of the month.

pub fn start_of_year(&self) -> Result<Self, DateTimeError>

Returns the start of the year for the DateTime instance.

Returns

Result<Self, DateTimeError> - A new DateTime instance at the start of the year.

pub fn end_of_year(&self) -> Result<Self, DateTimeError>

Returns the end of the year for the DateTime instance.

Returns

Result<Self, DateTimeError> - A new DateTime instance at the end of the year.

pub fn is_within_range(&self, start: &Self, end: &Self) -> bool

Checks if the DateTime instance falls within a specific range.

Arguments

start - The start DateTime of the range.

end - The end DateTime of the range.

Returns

bool - true if the DateTime falls within the range, otherwise false.

pub fn year(&self) -> i32

Returns the year of the DateTime instance.

pub fn month(&self) -> Month

Returns the month of the DateTime instance.

pub fn day(&self) -> u8

Returns the day of the month of the DateTime instance.

pub fn hour(&self) -> u8

Returns the hour of the DateTime instance.

pub fn minute(&self) -> u8

Returns the minute of the DateTime instance.

pub fn second(&self) -> u8

Returns the second of the DateTime instance.

pub fn microsecond(&self) -> u32

Returns the microsecond of the DateTime instance.

pub fn weekday(&self) -> Weekday

Returns the weekday of the DateTime instance.

pub fn ordinal(&self) -> u16

Returns the day of the year (ordinal) of the DateTime instance.

pub fn iso_week(&self) -> u8

Returns the ISO week number of the DateTime instance.

pub fn offset(&self) -> UtcOffset

Returns the UTC offset of the DateTime instance.

pub fn now() -> Self

Returns the current DateTime instance.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::now();
println!("Current time: {}", dt);

pub fn set_date( &self, year: i32, month: u8, day: u8, ) -> Result<Self, DateTimeError>

Sets a new date for the DateTime instance.

Arguments

year - The year to set.

month - The month to set (1-12).

day - The day of the month to set.

Returns

Result<Self, DateTimeError> - A new DateTime instance with the updated date, or an error if the date is invalid.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
let new_dt = dt.set_date(2024, 1, 1).expect("Setting date to 2024-01-01 should succeed");
println!("New date: {}", new_dt);

pub fn set_time( &self, hour: u8, minute: u8, second: u8, ) -> Result<Self, DateTimeError>

Sets a new time for the DateTime instance.

Arguments

hour - The hour to set (0-23).

minute - The minute to set (0-59).

second - The second to set (0-59).

Returns

Result<Self, DateTimeError> - A new DateTime instance with the updated time, or an error if the time is invalid.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
let new_dt = dt.set_time(12, 0, 0).expect("Setting time to 12:00:00 should succeed");
println!("New time: {}", new_dt);

pub fn is_valid_day(day: &str) -> bool

Validates whether the provided string represents a valid day of the month.

Arguments

day - A string slice representing the day (expected in “1” to “31”).

Returns

bool - Returns true if the day is valid, otherwise false.

Examples

use dtt::datetime::DateTime;
assert!(DateTime::is_valid_day("15"));
assert!(!DateTime::is_valid_day("32"));

pub fn is_valid_hour(hour: &str) -> bool

Validates whether the provided string represents a valid hour of the day.

Arguments

hour - A string slice representing the hour (expected in “0” to “23”).

Returns

bool - Returns true if the hour is valid, otherwise false.

Examples

use dtt::datetime::DateTime;
assert!(DateTime::is_valid_hour("23"));
assert!(!DateTime::is_valid_hour("24"));

pub fn is_valid_minute(minute: &str) -> bool

Validates whether the provided string represents a valid minute of the hour.

Arguments

minute - A string slice representing the minute (expected in “0” to “59”).

Returns

bool - Returns true if the minute is valid, otherwise false.

Examples

use dtt::datetime::DateTime;
assert!(DateTime::is_valid_minute("59"));
assert!(!DateTime::is_valid_minute("60"));

pub fn is_valid_second(second: &str) -> bool

Validates whether the provided string represents a valid second of the minute.

Arguments

second - A string slice representing the second (expected in “0” to “59”).

Returns

bool - Returns true if the second is valid, otherwise false.

Examples

use dtt::datetime::DateTime;
assert!(DateTime::is_valid_second("59"));
assert!(!DateTime::is_valid_second("60"));

pub fn is_valid_month(month: &str) -> bool

Validates whether the provided string represents a valid month of the year.

Arguments

month - A string slice representing the month (expected in “1” to “12”).

Returns

bool - Returns true if the month is valid, otherwise false.

Examples

use dtt::datetime::DateTime;
assert!(DateTime::is_valid_month("12"));
assert!(!DateTime::is_valid_month("13"));

pub fn is_valid_year(year: &str) -> bool

Validates whether the provided string represents a valid year.

Arguments

year - A string slice representing the year.

Returns

bool - Returns true if the year is valid, otherwise false.

Examples

use dtt::datetime::DateTime;
assert!(DateTime::is_valid_year("2023"));
assert!(!DateTime::is_valid_year("abcd"));

pub fn is_valid_microsecond(microsecond: &str) -> bool

Validates whether the provided string represents a valid microsecond.

Arguments

microsecond - A string slice representing the microsecond (expected in “0” to “999999”).

Returns

bool - Returns true if the microsecond is valid, otherwise false.

Examples

use dtt::datetime::DateTime;
assert!(DateTime::is_valid_microsecond("123456"));
assert!(!DateTime::is_valid_microsecond("1000000"));

pub fn is_valid_ordinal(ordinal: &str) -> bool

Validates whether the provided string represents a valid ordinal day of the year.

Arguments

ordinal - A string slice representing the ordinal day (expected in “1” to “366”).

Returns

bool - Returns true if the ordinal is valid, otherwise false.

Examples

use dtt::datetime::DateTime;
assert!(DateTime::is_valid_ordinal("200"));
assert!(!DateTime::is_valid_ordinal("400"));

pub fn is_valid_iso_week(week: &str) -> bool

Validates whether the provided string represents a valid ISO week number.

Arguments

week - A string slice representing the ISO week number (expected in “1” to “53”).

Returns

bool - Returns true if the ISO week number is valid, otherwise false.

Examples

use dtt::datetime::DateTime;
assert!(DateTime::is_valid_iso_week("10"));
assert!(!DateTime::is_valid_iso_week("54"));

pub fn is_valid_time(time: &str) -> bool

Validates whether the provided string represents a valid time in HH:MM:SS format.

Arguments

time - A string slice representing the time in HH:MM:SS format.

Returns

bool - Returns true if the time is valid, otherwise false.

Examples

use dtt::datetime::DateTime;
assert!(DateTime::is_valid_time("12:30:45"));
assert!(!DateTime::is_valid_time("24:00:00"));

pub fn is_valid_iso_8601(date: &str) -> bool

Validates whether the provided string is a valid ISO 8601 formatted date.

Arguments

date - A string slice representing the ISO 8601 date.

Returns

bool - Returns true if the date is valid, otherwise false.

Examples

use dtt::datetime::DateTime;
assert!(DateTime::is_valid_iso_8601("2023-01-01T12:00:00Z"));
assert!(!DateTime::is_valid_iso_8601("2023-13-01T12:00:00Z"));

pub fn add_months(&self, months: i32) -> Result<Self, DateTimeError>

Adds a specified number of months to the DateTime instance.

This method handles edge cases such as month-end dates and leap years. It also correctly handles negative month additions (subtractions).

Arguments

months - The number of months to add (can be negative for subtraction).

Returns

Result<Self, DateTimeError> - A new DateTime instance with the months added, or an error if the operation fails.

pub fn sub_months(&self, months: i32) -> Result<Self, DateTimeError>

Subtracts a specified number of months from the DateTime instance.

Arguments

months - The number of months to subtract.

Returns

Result<Self, DateTimeError> - A new DateTime instance with the months subtracted, or an error if the operation fails.

pub fn add_years(&self, years: i32) -> Result<Self, DateTimeError>

Adds a specified number of years to the DateTime instance.

This method handles leap years correctly.

Arguments

years - The number of years to add.

Returns

Result<Self, DateTimeError> - A new DateTime instance with the years added, or an error if the operation fails.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
let future_dt = dt.add_years(5).expect("Adding 5 years should succeed");
println!("Date after 5 years: {}", future_dt);

pub fn sub_years(&self, years: i32) -> Result<Self, DateTimeError>

Subtracts a specified number of years from the DateTime instance.

Arguments

years - The number of years to subtract.

Returns

Result<Self, DateTimeError> - A new DateTime instance with the years subtracted, or an error if the operation fails.

Examples

use dtt::datetime::DateTime;

let dt = DateTime::new();
let past_dt = dt.sub_years(5).expect("Subtracting 5 years should succeed");
println!("Date 5 years ago: {}", past_dt);

pub fn format_time_in_timezone( tz: &str, format: &str, ) -> Result<String, DateTimeError>

Formats the current time for a specific timezone.

Arguments

tz - A string slice that holds the timezone abbreviation (e.g., “UTC”, “EST”, “PST”).

format - A string slice that specifies the desired output format.

Returns

Result<String, DateTimeError> - The formatted time string or an error if the operation fails.

Examples

use dtt::datetime::DateTime;
use time::format_description;

let formatted_time = DateTime::format_time_in_timezone("PST", "[hour repr:12]:[minute] [period]").expect("Formatting failed");
println!("Current time in Los Angeles: {}", formatted_time);